break parser out into a separate package
