+++ resource/generation/options.go
@@ -206,5 +206,22 @@ func defaultTypescriptOverrides() map[string]string {
 		reflect.TypeOf(decimal.Decimal{}).String():     "number",
 		reflect.TypeOf(decimal.NullDecimal{}).String(): "number",
 		reflect.TypeOf(time.Time{}).String():           "Date",
+		"bool":                                         "boolean",
+		"string":                                       "string",
+		"int":                                          "number",
+		"int8":                                         "number",
+		"int16":                                        "number",
+		"int32":                                        "number",
+		"int64":                                        "number",
+		"uint":                                         "number",
+		"uint8":                                        "number",
+		"uint16":                                       "number",
+		"uint32":                                       "number",
+		"uint64":                                       "number",
+		"uintptr":                                      "number",
+		"float32":                                      "number",
+		"float64":                                      "number",
+		"complex64":                                    "number",
+		"complex128":                                   "number",
 	}
 }
