+++ resource/generation/parser/test_helpers.go
@@ -1,0 +1,8 @@
+package parser
+
+import "testing"
+
+func NewTestType(t *testing.T, name string) Type {
+	t.Helper()
+	return Type{name: name}
+}
